Refactoring and fixes (#36)
Version 1.32 (January 23, 2021):
  - (Atari2.0) Fixes spritetool_clean.asm not being in the asm/ folder.
  - (Atari2.0) Now permits nested shared routine calls, fixing https://github.com/JackTheSpades/SpriteToolSuperDelux/issues/28
  - (Atari2.0) Updated asar.dll to 1.80, fixes https://github.com/JackTheSpades/SpriteToolSuperDelux/issues/35
  - (Atari2.0) Updated and fixed make, Makefile and .gitignore
  - (Atari2.0) Added escaping to ! in filepaths, fixes https://github.com/JackTheSpades/SpriteToolSuperDelux/issues/16
  - (Atari2.0) Added autoclean to all sprite pointers, fixes https://github.com/JackTheSpades/SpriteToolSuperDelux/issues/34
  - (Atari2.0) Fixed various compiler warnings, swapped fread arguments causing an error when writing extmod file.
  - (Atari2.0) Fixed snestopc not working correctly on SA-1 roms bigger than 2 mbs
  - (Atari2.0) Greatly refactored and reworked MeiMei, embedded both ROM classes in one
  - (Atari2.0) Made Pixi work on +6mb ROMs correctly
  - (Atari2.0) Now can use -nr to specify max number of shared routines
  - (Atari2.0) Fixed various memory leaks, along with a couple of asar warnings related to relative paths being used.
  - (Atari2.0) You can now use comments in your list.txt file
  - (Atari2.0) Completely rewritten populate_sprite_list and relative_directory to fix various bugs related to pathing.

+++ src/file_io.cpp
@@ -10,9 +10,9 @@ FILE *open(const char *name, const char *mode) {
 	return file;
 }
 
-int file_size(FILE *file) {
+size_t file_size(FILE *file) {
 	fseek(file, 0, SEEK_END);
-	int size = ftell(file);
+	size_t size = ftell(file);
 	fseek(file, 0, SEEK_SET);
 	return size;
 }
